requirements.txt
fastapi
sqlalchemy
pydantic
psycopg2
python-dotenv